Update the integration to use uint64_t time values to get an accurate time step
Before there were small rounding issues converting from uint64_t to double, then differencing doubles.  Integrating with a slightly wrong value can cause drift in the integrated states.  Now the integration time step is computed precisely using uint64_5 integers to an accuracy of nano-seconds.

+++ src/simulation/dynamics/spacecraftSystem/spacecraftSystem.h
@@ -204,8 +204,8 @@ public:
     void attachSpacecraftToPrimary(SpacecraftUnit *newSpacecraft, std::string dockingPortNameOfNewSpacecraft, std::string dockingToPortName);  //!< Attaches a spacecraft to the primary spacecraft chain
     void addSpacecraftUndocked(SpacecraftUnit *newSpacecraft);  //!< Attaches a spacecraft to the primary spacecraft chain
     void determineAttachedSCStates();  //!< class method
-    void preIntegration(double callTime) final;  //!< pre-integration steps
-    void postIntegration(double callTime) final;  //!< post-integration steps
+    void preIntegration(uint64_t callTime) final;  //!< pre-integration steps
+    void postIntegration(uint64_t callTime) final;  //!< post-integration steps
 
 private:
     Eigen::MatrixXd *sysTime;            //!< [s] System time
